Removed redundant components from Bundle and Robot
